// src/api/generated/model/getPrivateSubmissionsResponseItemDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { GetPrivateSubmissionsResponseItemDtoStatus } from './getPrivateSubmissionsResponseItemDtoStatus';

export interface GetPrivateSubmissionsResponseItemDto {
  /** Unique identifier */
  id: string;
  /** Timestamp of when the record was created */
  createdAt: string;
  /** Timestamp of the last update of the record */
  updatedAt: string;
  /** Собственные тексты креатива от амбассадора. Разрешены только если allowAmbassadorText=true у задачи. */
  texts?: string[];
  /** ID собственных медиафайлов амбассадора. Разрешены только если allowAmbassadorMedia=true у задачи. */
  mediaFileIds?: string[];
  /** Целевые ссылки перехода креатива (target_urls в ОРД) */
  targetUrls?: string[];
  /**
   * Комментарий от амбассадора
   * @nullable
   */
  comment?: string | null;
  /** Статус проверки ответа */
  status: GetPrivateSubmissionsResponseItemDtoStatus;
  /**
   * Комментарий от проверяющего
   * @nullable
   */
  reviewComment?: string | null;
  /**
   * Маркер рекламы (erid), появляется после регистрации креатива в ОРД
   * @nullable
   */
  erid?: string | null;
  /** ID приватного задания */
  privateTaskId: string;
  /** ID амбассадора */
  ambassadorId: string;
}